refactor: move test db logic to separate package
In internal/river tests we imported `"github.com/canonical/jimm/v3/internal/testutils/jimmtest"` for creation of test DBs. That package imports internal/jimm.
Internal/jimm will (optionally) import internal/river if we have a wrapper around River's client. This change ensures DB creation is not wrapped up in this dependency tree.
