+++ internal/jimm/auditlog/cleanup_test.go
@@ -14,6 +14,7 @@ import (
 	"github.com/canonical/jimm/v3/internal/dbmodel"
 	"github.com/canonical/jimm/v3/internal/jimm/auditlog"
 	"github.com/canonical/jimm/v3/internal/testutils/jimmtest"
+	"github.com/canonical/jimm/v3/internal/testutils/testdb"
 )
 
 func TestAuditLogCleanupServicePurgesLogs(t *testing.T) {
@@ -23,7 +24,7 @@ func TestAuditLogCleanupServicePurgesLogs(t *testing.T) {
 	ctx := context.Background()
 
 	db := &db.Database{
-		DB: jimmtest.PostgresDB(c, time.Now),
+		DB: testdb.PostgresDB(c, time.Now),
 	}
 	err := db.Migrate(context.Background())
 	c.Assert(err, qt.IsNil)
