+++ internal/jimm/sshkeys/sshkeys_test.go
@@ -19,6 +19,7 @@ import (
 	"github.com/canonical/jimm/v3/internal/jimm/sshkeys"
 	"github.com/canonical/jimm/v3/internal/openfga"
 	"github.com/canonical/jimm/v3/internal/testutils/jimmtest"
+	"github.com/canonical/jimm/v3/internal/testutils/testdb"
 )
 
 type sshKeysManagerSuite struct {
@@ -34,7 +35,7 @@ func (s *sshKeysManagerSuite) Init(c *qt.C) {
 	ctx := context.Background()
 	// Setup DB
 	db := &db.Database{
-		DB: jimmtest.PostgresDB(c, time.Now),
+		DB: testdb.PostgresDB(c, time.Now),
 	}
 	err := db.Migrate(context.Background())
 	c.Assert(err, qt.IsNil)
