+++ internal/river/utils_test.go
@@ -7,24 +7,27 @@ import (
 	"time"
 
 	"github.com/canonical/jimm/v3/internal/db"
-	"github.com/canonical/jimm/v3/internal/testutils/jimmtest"
+	"github.com/canonical/jimm/v3/internal/testutils/testdb"
 	qt "github.com/frankban/quicktest"
 	"github.com/riverqueue/river/rivertype"
 )
 
 func setupTestDB(c *qt.C) (*db.Database, *sql.DB) {
 	db := &db.Database{
-		DB: jimmtest.PostgresDB(c, time.Now),
+		DB: testdb.PostgresDB(c, time.Now),
 	}
 	err := db.Migrate(c.Context())
 	c.Assert(err, qt.IsNil)
+
 	err = MigrateRiver(c.Context(), db)
 	c.Assert(err, qt.IsNil)
+
 	sqlDB, err := db.SqlDB()
 	c.Assert(err, qt.IsNil)
 	c.Cleanup(func() {
 		c.Check(sqlDB.Close(), qt.IsNil)
 	})
+
 	return db, sqlDB
 }
